Refactor: Role 위치 변경.
