+++ src/main/java/com/sprint/mission/discodeit/dto/request/RoleUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.dto.request;
 
-import com.sprint.mission.discodeit.security.Role;
+import com.sprint.mission.discodeit.entity.Role;
 import java.util.UUID;
 
 public record RoleUpdateRequest(
