+++ src/main/java/com/sprint/mission/discodeit/entity/Role.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.security;
+package com.sprint.mission.discodeit.entity;
 
 public enum Role {
   ADMIN,
